tesseract-api/requirements.txt
blinker==1.7.0
click==8.1.7
Flask==3.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
packaging==23.2
Pillow==10.1.0
PyPDF2==3.0.1
pytesseract==0.3.10
Werkzeug==3.0.1
